fixed error code

+++ api/src/app/services/orders/deleteManyOrdersService.ts
@@ -1,11 +1,13 @@
 import { Order } from '../../../models/Order.js';
 
 export const deleteManyOrdersService = async () => {
-  const result = await Order.deleteMany();
+  const ordersExists = await Order.exists({});
 
-  if (result.deletedCount == 0) {
-    throw new Error('No orders available');
+  if (!ordersExists) {
+    throw new Error('No orders availables.');
   }
+
+  const result = await Order.deleteMany();
 
   return { message: 'Orders deleted successfully', deletedCount: result.deletedCount };
 };
